requirements.txt
fastapi
uvicorn[standard]
fastmcp
python-dotenv
pyjwt[crypto]
httpx
